Update cards by Terrie
Updated a.o. domain
Fixed bug with a.o. spawnpoint
Fixed bug with a.o. nexus
Changed the implementation of a.o. nexus (so that it's easier to play with but won't work with cards like soul exchange or lair of darkness for now)
Removed the bug test version of a.o. nexus