Update cards by Somen
A lua table in which every banned card is a key with the value true. It goes in expansions/script. I will be updating some cards so that they reference this soon. Other people can feel free to edit this file when the banned list changes since I might not notice right away
* Edited to auto-update by Lyris
I moved my sounds to a seperate zip file
Update including fixing a few create effects